+++ middleware.tsx
@@ -16,6 +16,7 @@ export const config = {
     "/collections/:path*",
     "/divisions",
     "/divisions/:path*",
+    "/items/:path*",
     "/search/:path*",
   ],
 };